move wallet tooling in separate module to ease reusability
